+++ ServerStarter/src/main/java/ca/qc/johnabbott/cs616/server/DrugRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
  * Created by Gabriel Charlebois on 2017-12-10.
  */
 @RepositoryRestResource(collectionResourceRel = "drug", path = "drug")
-public interface DrugRepository  extends CrudRepository<Drug,Long> {
+public interface DrugRepository  extends CrudRepository<Drug,Integer> {
 }
